Update Routes web

+++ routes/web.php
@@ -15,13 +15,14 @@ use Illuminate\Support\Facades\Route;
 
 
 
-Route::get('posts', 'Admin\ComicController@index');
-Route::get('posts/create', 'Admin\ComicController@create');
-Route::post('posts', 'Admin\ComicController@store');
-Route::get('posts/{post}', 'Admin\ComicController@show');
-Route::get('posts/{post}/edit', 'Admin\ComicController@edit');
-Route::put('posts/{post}', 'Admin\ComicController@update');
-Route::delete('posts/{post}', 'Admin\ComicController@destroy');
+Route::get('posts', 'Admin\ComicController@index')->name('admin.comic.index');
+Route::get('posts/create', 'Admin\ComicController@create')->name('admin.comic.create');
+Route::post('posts', 'Admin\ComicController@store')->name('admin.comic.store');
+
+Route::get('posts/{post}', 'Admin\ComicController@show')->name('admin.comic.show');
+Route::get('posts/{post}/edit', 'Admin\ComicController@edit')->name('admin.comic.edit');
+Route::put('posts/{post}', 'Admin\ComicController@update')->name('admin.comic.update');
+Route::delete('posts/{post}', 'Admin\ComicController@destroy')->name('admin.comic.destroy');
